Structure public and private pages ensuring default authentication
* Before this change we had to remember to add the `Secured` module to every controller. This is easy to forget. We want to make sure most pages are authenticated by default and allow health checks and public pages to be clearly skipped.
* We know there will be more user groups later so it felt too premature to start addeding `administrator` folder structures. I imagine these will come in good time and be made easier by this change. It may be extended as so: `users/administrators/users_controller.rb`, and `users/delivery_partners/projects_controller.rb`
* Removed un anunsed about page view and test
