--- app/controllers/dashboards_controller.rb
@@ -1,6 +1,0 @@
-class DashboardsController < ApplicationController
-  include Secured
-
-  def show
-  end
-end
