+++ app/controllers/public/visitors_controller.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class Public::VisitorsController < Public::BaseController
+end
