+++ app/controllers/users/dashboards_controller.rb
@@ -1,0 +1,4 @@
+class Users::DashboardsController < Users::BaseController
+  def show
+  end
+end
