+++ app/views/users/dashboards/show.html.haml
