+++ spec/features/public/visitors/home_page_spec.rb
