+++ spec/features/users/users_can_sign_out_spec.rb
